backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
livekit-agents[openai,deepgram,silero]>=0.11.0
livekit>=0.11.0
openai>=1.55.0
python-dotenv>=1.0.1
aiosqlite>=0.20.0
